app/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv

from app import auth, models
from app.database import engine
from app.routes import auth as auth_routes
from app.routes import tasks as task_routes


load_dotenv()
auth.validate_auth_settings()
models.Base.metadata.create_all(bind=engine)

app = FastAPI(title="Task Manager API", version="1.0.0")

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(auth_routes.router)
app.include_router(task_routes.router)


@app.get("/health")
def health_check():
    return {"status": "ok"}
